Break out util into respective files, refocus on hoggerstate deserialization
