+++ hogger/util/__init__.py
@@ -1,16 +1,34 @@
-from .misc import Duration, Money
+# from .misc import Duration, Money
+# from .misc import LookupID, Lookup, Money, Duration
+# from .utils import EnumMapUtils, EnumUtils, IntFlagUtils
+
+from .duration import Duration
+from .enum import EnumUtils
+from .enummap import EnumMapUtils
 from .errors import InvalidValueException
-from .misc import LookupID, Lookup, Money, Duration
-from .utils import EnumMapUtils, EnumUtils, IntFlagUtils
+from .intflag import IntFlagUtils
+from .lookup import Lookup, LookupID
+from .money import Money
+from .utils import pydantic_annotation
+
+
 
 __all__ = [
-    # misc
+    # duration
     "Duration",
-    "Money",
-    "LookupID",
-    "Lookup",
-    # utils
-    "EnumMapUtils",
+    # enum
     "EnumUtils",
+    # enummap
+    "EnumMapUtils",
+    # errors
+    "InvalidValueException",
+    # intflag
     "IntFlagUtils",
+    # lookup
+    "Lookup",
+    "LookupID",
+    # money
+    "Money",
+    # utils
+    "pydantic_annotation"
 ]
